% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_contract.R
\name{is_artifact_contract}
\alias{is_artifact_contract}
\title{Test whether an object is an Artifact Contract}
\usage{
is_artifact_contract(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"artifact_contract"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is an Artifact Contract
}
