% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_contract.R
\name{is_dependency_contract}
\alias{is_dependency_contract}
\title{Test whether an object is a Dependency Contract}
\usage{
is_dependency_contract(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"dependency_contract"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a Dependency Contract
}
