% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_contract.R
\name{is_interface_contract}
\alias{is_interface_contract}
\title{Test whether an object is an Interface Contract}
\usage{
is_interface_contract(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"interface_contract"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is an Interface Contract
}
