% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification_contract.R
\name{is_verification_contract}
\alias{is_verification_contract}
\title{Test whether an object is a Verification Contract}
\usage{
is_verification_contract(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"verification_contract"},
\code{FALSE} otherwise.
}
\description{
Test whether an object is a Verification Contract
}
