% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_contract.R
\name{print.artifact_contract}
\alias{print.artifact_contract}
\title{Print an Artifact Contract}
\usage{
\method{print}{artifact_contract}(x, ...)
}
\arguments{
\item{x}{An \code{artifact_contract} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Artifact Contract
}
