% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_selection.R
\name{print.candidate_selection_object}
\alias{print.candidate_selection_object}
\title{Print a Candidate Selection Object}
\usage{
\method{print}{candidate_selection_object}(x, ...)
}
\arguments{
\item{x}{A \code{candidate_selection_object} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Candidate Selection Object
}
