% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_contract.R
\name{print.dependency_contract}
\alias{print.dependency_contract}
\title{Print a Dependency Contract}
\usage{
\method{print}{dependency_contract}(x, ...)
}
\arguments{
\item{x}{A \code{dependency_contract} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Dependency Contract
}
