% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_contract.R
\name{print.interface_contract}
\alias{print.interface_contract}
\title{Print an Interface Contract}
\usage{
\method{print}{interface_contract}(x, ...)
}
\arguments{
\item{x}{An \code{interface_contract} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Interface Contract
}
