% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{print.interpretation_object}
\alias{print.interpretation_object}
\title{Print an Interpretation Object}
\usage{
\method{print}{interpretation_object}(x, ...)
}
\arguments{
\item{x}{An \code{interpretation_object} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Interpretation Object
}
